# openapi/paths.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification. You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8000
...
paths:
...
  /users/me:
    get:
      tags:
        - users
      summary: Lấy thông tin user
      description: Lấy thông tin của người dùng hiện tại
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lấy thông tin user thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - users
      summary: Cập nhật thông tin user
      description: Cập nhật thông tin của người dùng hiện tại
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin cập nhật user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeBody'
        required: true
      responses:
        '200':
          description: Cập nhật thông tin user thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/verify-email:
    post:
      tags:
        - users
      summary: Xác thực email
      description: Xác thực tài khoản của người dùng bằng token được gửi qua email
      operationId: verifyEmail
      requestBody:
        description: Thông tin để xác thực email
        content:
          application/json:
            schema:
              type: object
              required:
                - email_verify_token
              properties:
                email_verify_token:
                  type: string
                  description: Token xác thực email được gửi qua email
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MiwidmVyaWZ5IjowLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTY5MTQ2OTYyNn0.XYZabc123...
        required: true
      responses:
        '200':
          description: Xác thực email thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad Request - Invalid token or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/refresh-token:
    post:
      tags:
        - users
      summary: Lấy lại access token
      description: Làm mới access token bằng refresh token
      operationId: refreshToken
      requestBody:
        description: Thông tin làm mới access token
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
        required: true
      responses:
        '200':
          description: Làm mới access token thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized - Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/resend-verify-email:
    post:
      tags:
        - users
      summary: Gửi lại email xác thực
      description: Gửi lại email để xác thực tài khoản
      operationId: resendVerifyEmail
      requestBody:
        description: Thông tin gửi lại email xác thực
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  example: duthanhduoc100@gmail.com
        required: true
      responses:
        '200':
          description: Gửi lại email xác thực thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad Request - Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/forgot-password:
    post:
      tags:
        - users
      summary: Quên mật khẩu
      description: Gửi email để đặt lại mật khẩu
      operationId: forgotPassword
      requestBody:
        description: Thông tin để gửi email đặt lại mật khẩu
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  example: duthanhduoc100@gmail.com
        required: true
      responses:
        '200':
          description: Nhận được email reset password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Bad Request - Invalid email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/reset-password:
    post:
      tags:
        - users
      summary: Reset mật khẩu
      description: Đặt lại mật khẩu bằng token
      operationId: resetPassword
      requestBody:
        description: Thông tin để đặt lại mật khẩu
        content:
          application/json:
            schema:
              type: object
              required:
                - forgot_password_token
                - password
                - confirm_password
              properties:
                forgot_password_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
                password:
                  type: string
                  example: Duoc123!
                confirm_password:
                  type: string
                  example: Duoc123!
        required: true
      responses:
        '200':
          description: Reset mật khẩu thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized - Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/change-password:
    put:
      tags:
        - users
      summary: Đổi mật khẩu
      description: Đổi mật khẩu của người dùng hiện tại
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin để đổi mật khẩu
        content:
          application/json:
            schema:
              type: object
              required:
                - old_password
                - password
                - confirm_password
              properties:
                old_password:
                  type: string
                  example: Duoc123!
                password:
                  type: string
                  example: Duoc1234!
                confirm_password:
                  type: string
                  example: Duoc1234!
        required: true
      responses:
        '200':
          description: Đổi mật khẩu thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /medias/upload-image:
    post:
      tags:
        - medias
      summary: Upload ảnh
      description: Upload một hoặc nhiều ảnh
      operationId: uploadImage
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload ảnh
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 4
        required: true
      responses:
        '200':
          description: Upload ảnh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload ảnh thành công
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Media'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /medias/upload-video:
    post:
      tags:
        - medias
      summary: Upload video
      description: Upload một video
      operationId: uploadVideo
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload video
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
        required: true
      responses:
        '200':
          description: Upload video thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload video thành công
                  result:
                    $ref: '#/components/schemas/Media'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: 64be0ad2e43d2464394feedb
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: johndoe@example.com
        date_of_birth:
          type: string
          format: date-time
          example: 2023-06-08T10:17:31.096Z
        created_at:
          type: string
          format: date-time
          example: 2023-03-08T12:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2023-03-08T12:00:00Z
        verify:
          $ref: '#/components/schemas/UserVerifyStatus'
        twitter_circle:
          type: array
          items:
            type: string
            format: MongoId
          example: ['64be0ad2e43d2464394feedb', '64be0ad2e43d2464394feedc']
        bio:
          type: string
          example: This is my bio.
        location:
          type: string
          example: San Francisco, CA
        website:
          type: string
          example: https://www.example.com
        username:
          type: string
          example: johndoe
        avatar:
          type: string
          example: http://localhost:4000/images/avatars/johndoe.jpg
        cover_photo:
          type: string
          example: http://localhost:4000/images/covers/johndoe.jpg
    UserVerifyStatus:
      type: number
      enum:
        - 0 # Unverified
        - 1 # Verified
        - 2 # Banned
      example: 1
...
    UpdateMeBody:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        date_of_birth:
          type: string
          format: date-time
          example: 2023-06-08T10:17:31.096Z
        bio:
          type: string
          example: This is my bio.
        location:
          type: string
          example: San Francisco, CA
        website:
          type: string
          example: https://www.example.com
        username:
          type: string
          example: johndoe
        avatar:
          type: string
          example: http://localhost:4000/images/avatars/johndoe.jpg
        cover_photo:
          type: string
          example: http://localhost:4000/images/covers/johndoe.jpg
...
    MediaType:
      type: number
      enum:
        - 0 # Image
        - 1 # Video
        - 2 # HLS
      example: 0
    Media:
      type: object
      properties:
        url:
          type: string
          example: http://localhost:4000/images/avatars/johndoe.jpg
        type:
          $ref: '#/components/schemas/MediaType'
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: Operation successful
        result:
          type: object
          additionalProperties: true
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: An error occurred
        code:
          type: string
          example: ERROR_CODE
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT